refactor(context,llm): remove dead field and deduplicate build_tool_description
Remove fidelity_config from ContextAssemblyInput: the field was always
None at every call site and never read by assembler.rs — fidelity
scoring happens after gather() returns in the prepare_context pipeline.

Deduplicate build_tool_description from claude and openai backends into
a new shared module (crates/zeph-llm/src/tool_desc.rs). Both backends
now import the single implementation with a shared static WARNED guard
for once-per-session schema-overflow warnings.

Closes #4595
Closes #4577
